Introduce fn make_transports in test harness
