+++ tests/harness/mod.rs
@@ -1,3 +1,5 @@
+mod transport;
+
 use bitcoin::{util::psbt::PartiallySignedTransaction, Address, Amount};
 use bitcoin_harness::{bitcoind_rpc::PsbtBase64, Bitcoind};
 use reqwest::Url;
@@ -5,6 +7,8 @@ use thor::{
     create::{BuildFundingPSBT, SignFundingPSBT},
     BroadcastSignedTransaction, NewAddress,
 };
+
+pub use transport::make_transports;
 
 pub struct Wallet(pub bitcoin_harness::Wallet);
